requirements.txt
playwright==1.41.2
requests==2.31.0
beautifulsoup4==4.12.3
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pypdf==4.0.1
python-docx==1.1.0
